Improve example for `rtype:periodic`
If you are very late buying birthday flowers, it might be better to wait
until next year (depends on the person, I guess). But rent is more
black-and-white: it is due regularly, and if not paid then the
obligations accumulate.

The replacement example was inspired by one of the examples given at
https://orgmode.org/manual/Repeated-tasks.html