Fix "Filter Precedence" page explanation of grouping of report and CLI filters
The documentation originally reported that parenthesis should be used to
group CLI filters and report filters properly, but this is not the case.

Particularly, the docs used to say that:

```bash
$ task project:Home or project:Garden list
```

is interpreted as:

`status:pending and project:Home or project:Garden`

instead of:

`status:pending and (project:Home or project:Garden)`

However, that's not the case. This can be shown by running:

```bash
$ task rc.verbose=yes rc.report.ls.filter="status:pending" project:Home or project:Garden ls
...
Filter: ( status = pending ) and ( project = Home or project = Garden )
...
```

Because of this, the documentation was updated. Also, another example
was used to continue explaining the use of parentheses.