Add a `tw-release` script (#1269)
This does everything in the RELEASING.md file in the Taskwarrior repo,
as well as updating manpages.

It turns out we never use the `tarball` property in releases, so that is
removed. We also never link to the tarballs committed to this repo, so
this no longer updates them -- but leaves the existing tarballs in case
other people are using those links.